requirements.txt
alabaster==0.7.12
arrow==1.2.3
astroid==2.14.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==23.1.0
autopep8==2.0.4
Babel==2.11.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==24.3.0
bleach==4.1.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.8
docstring-to-markdown==0.11
docutils==0.18.1
executing==0.8.3
fastjsonschema==2.16.2
flake8==7.0.0
fonttools==4.51.0
idna==3.7
imagesize==1.4.1
importlib-metadata==7.0.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.28.0
ipython==8.20.0
isort==5.9.3
jaraco.classes==3.2.1
jedi==0.18.1
jellyfish==1.0.1
Jinja2==3.1.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-pygments==0.2.2
keyring==24.3.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.0
mistune==2.0.4
more-itertools==10.1.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
numpydoc==1.5.0
packaging==23.2
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.15.1
pylint==2.16.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-slugify==5.0.2
pytoolconfig==1.2.6
pytz==2024.1
pywin32==305.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
rich==13.3.5
rope==1.12.0
rpds-py==0.10.6
Rtree==1.0.1
setuptools==69.5.1
sip==6.7.12
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.5.1
spyder-kernels==2.5.0
stack-data==0.2.0
text-unidecode==1.3
textdistance==4.2.1
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.1
tornado==6.3.3
traitlets==5.7.1
ujson==5.4.0
Unidecode==1.2.0
urllib3==2.2.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
whatthepatch==1.0.2
wheel==0.43.0
win-inet-pton==1.1.0
wrapt==1.14.1
yapf==0.40.2
zipp==3.17.0
